Enable authenticated Exports
It is now possible to export restricted namespaces and pages as long as
you can authenticate with a user that has DELETE access to the wiki.
The credentials will be forwarded the first time a request is made (if
not exporting via AJAX) so all the subsequent requests are
authenticated (via Cookie).