Add default authentication
This adds default authentication for exports using the export URL. If
you try to export pages that would not be authenticated, you can set a
user (together with the password) in the settings. These credentials
will the be used for the export of pages that would not be
authenticated otherwise.