Add option to ignore errors during export via GUI
Using the export via the GUI will result in not delivering the export
result while there are log messages with errors. In rare  cases this
might still be the desired case. If so, you can enable the “Do not
interrupt export with errors” setting in the global admin settings
panel.